Moved base actions to another namespace and refactored using statements
