+++ src/Pokegraf.Application.Contract/BotActions/Common/ICommandAction.cs
@@ -1,4 +1,4 @@
-namespace Pokegraf.Application.Contract.Common.Actions
+namespace Pokegraf.Application.Contract.BotActions.Common
 {
     public interface ICommandAction : IBotAction
     {
