+++ src/Pokegraf.Application.Contract/Common/Strategy/IStrategyContext.cs
@@ -1,6 +1,5 @@
 using System.Collections.Generic;
 using Pokegraf.Application.Contract.BotActions.Common;
-using Pokegraf.Application.Contract.Common.Actions;
 
 namespace Pokegraf.Application.Contract.Common.Strategy
 {
