+++ src/Pokegraf.Application.Implementation/BotActions/Callbacks/PokemonNext/PokemonNextCallbackAction.cs
@@ -1,6 +1,5 @@
 using Pokegraf.Application.Contract.Common.Context;
 using Pokegraf.Application.Implementation.BotActions.Common;
-using Pokegraf.Application.Implementation.Common.Actions;
 
 namespace Pokegraf.Application.Implementation.BotActions.Callbacks.PokemonNext
 {
