+++ src/Pokegraf.Application.Implementation/Common/Strategy/StrategyContext.cs
@@ -1,5 +1,5 @@
 using System.Collections.Generic;
-using Pokegraf.Application.Contract.Common.Actions;
+using Pokegraf.Application.Contract.BotActions.Common;
 using Pokegraf.Application.Contract.Common.Strategy;
 
 namespace Pokegraf.Application.Implementation.Common.Strategy
