+++ src/Pokegraf.Application.Implementation/Service/TelegramService.cs
@@ -4,10 +4,9 @@ using System.Threading.Tasks;
 using MediatR;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
-using Pokegraf.Application.Contract.BotActions.Common;
-using Pokegraf.Application.Contract.Common.Actions;
 using Pokegraf.Application.Contract.Common.Client;
 using Pokegraf.Application.Contract.Common.Context;
+using Pokegraf.Application.Contract.Common.Strategy;
 using Pokegraf.Application.Contract.Service;
 using Telegram.Bot.Args;
 using Telegram.Bot.Types.Enums;
